==> Analysis/02_main/05_plots_main.Rmd <==
---
title: "05_plots"
output: html_document
---
```{r,warning=FALSE}
# load required library
library(tidyverse)
```



```{r, message=FALSE}
# load the dataset
cleaned_data <- read_csv('../../Data/02_main/02_cleaned_main_data.csv')

cleaned_data
```

# Plotting the Participants' Responses per Condition and Dilemma Type
```{r, message=FALSE}
response_counts_overall <- cleaned_data %>% group_by(ingroup_norm, both_infos, dilemma_rating)%>%
    summarize(count = length(dilemma_rating)) %>% ungroup()
max_response_count_overall <- max(response_counts_overall$count)

response_counts <- cleaned_data %>% group_by(ingroup_norm, both_infos, dilemma_rating, dilemma_type)%>%
    summarize(count = length(dilemma_rating)) %>% ungroup()
max_response_count <- max(response_counts$count)

response_counts_real <- response_counts %>% filter(dilemma_type == 'real_dilemma') %>% select(-dilemma_type)
response_counts_philo <- response_counts %>% filter(dilemma_type == 'philosophical_dilemma') %>% select(-dilemma_type)
```

```{r, message=FALSE}
vertical_bar_width = 1
horizontal_bar_width = 0.4

props_overall <- cleaned_data %>% group_by(ingroup_norm, both_infos, dilemma_rating) %>% 
  summarize(
    count = length(dilemma_rating),
    relative_proportion = count/max_response_count_overall,
    xmin = ingroup_norm[1] - vertical_bar_width/1.5,
    xmax = xmin + relative_proportion*vertical_bar_width,
    ymin = dilemma_rating[1] - horizontal_bar_width, 
    ymax = dilemma_rating[1] + horizontal_bar_width
  ) %>% ungroup()
props_overall$both_infos = factor(props_overall$both_infos, labels=c("Only ingroup norm shown", "Both norms shown"))

props <- cleaned_data %>% group_by(ingroup_norm, both_infos, dilemma_rating, dilemma_type) %>% 
  summarize(
    count = length(dilemma_rating),
    relative_proportion = count/max_response_count,
    xmin = ingroup_norm[1] - vertical_bar_width/1.5,
    xmax = xmin + relative_proportion*vertical_bar_width,
    ymin = dilemma_rating[1] - horizontal_bar_width, 
    ymax = dilemma_rating[1] + horizontal_bar_width
  ) %>% ungroup()
props$both_infos = factor(props$both_infos, labels=c("Only ingroup norm shown", "Both norms shown"))

props_real <- props %>% filter(dilemma_type == 'real_dilemma') %>% select(-dilemma_type)
props_philo <- props %>% filter(dilemma_type == 'philosophical_dilemma') %>% select(-dilemma_type)

means_overall <- cleaned_data %>% group_by(ingroup_norm, both_infos) %>%
    summarize(meanResponse = mean(dilemma_rating),
              stdErrResponse = sd(dilemma_rating)/sqrt(length(dilemma_rating)),
              boundLower = meanResponse - stdErrResponse,
              boundUpper = meanResponse + stdErrResponse
    ) %>% ungroup()
means_overall$both_infos = factor(means_overall$both_infos, labels=c("Only ingroup norm shown", "Both norms shown"))

means <- cleaned_data %>% group_by(ingroup_norm, both_infos, dilemma_type) %>%
    summarize(meanResponse = mean(dilemma_rating),
              stdErrResponse = sd(dilemma_rating)/sqrt(length(dilemma_rating)),
              boundLower = meanResponse - stdErrResponse,
              boundUpper = meanResponse + stdErrResponse
    ) %>% ungroup()
means$both_infos = factor(means$both_infos, labels=c("Only ingroup norm shown", "Both norms shown"))

means_real <- means %>% filter(dilemma_type == 'real_dilemma') %>% select(-dilemma_type)
means_philo <- means %>% filter(dilemma_type == 'philosophical_dilemma') %>% select(-dilemma_type)
```
## For Both Dilemmas
```{r}
library(scales)
dilemma_decision_plot <- ggplot()+
    geom_col(data=means_overall, aes(x=ingroup_norm, y=meanResponse, fill=factor(ingroup_norm)), width=vertical_bar_width, alpha=0.5) +
    geom_rect(data=props_overall, aes(xmin=xmin, ymin=ymin, xmax=xmax, ymax=ymax, group=ingroup_norm), fill = "#9c9d9e", alpha = 0.8) +
    scale_y_continuous(name="Preference", limits=c(1,6), oob = rescale_none, breaks = 1:6, labels=c("Definitely \n Active Option", "","","","", "Definitely \n Passive Option")) + 
    scale_x_continuous(name = "Ingroup Norm", breaks=c(-1,1), labels=c("Active Option", "Passive Option"))+
    scale_fill_manual(breaks = c(-1, 1), labels=c("Active Option", "Passive Option"), values = c("#3148f7", "#35b5ff")) +
    facet_wrap(~both_infos) + 
    scale_alpha(range = c(0, 1)) +
    theme(panel.spacing = unit(4, "lines"),
          strip.text = element_text(size = 14),
          panel.spacing.x = unit(3, "cm")) +
    guides(fill=FALSE, group=FALSE, alpha=FALSE) +
  ggtitle('Responses to Both Dilemma Types') +
  theme(plot.title = element_text(hjust = 0.5))
dilemma_decision_plot
```
## For the Realistic Dilemma
```{r}
dilemma_decision_plot_real <- ggplot()+
    geom_col(data=means_real, aes(x=ingroup_norm, y=meanResponse, fill=factor(ingroup_norm)), width=vertical_bar_width, alpha=0.5) +
    geom_rect(data=props_real, aes(xmin=xmin, ymin=ymin, xmax=xmax, ymax=ymax, group=ingroup_norm), fill = "#9c9d9e", alpha = 0.8) +
    scale_y_continuous(name="Preference", limits=c(1,6), oob = rescale_none, breaks = 1:6, labels=c("Definitely \n Reporting To Police", "","","","", "Definitely \n Not Reporting \n To Police")) + 
    scale_x_continuous(name = "Ingroup Norm", breaks=c(-1,1), labels=c("Report \n To Police", "Not Report \n To Police"))+
    scale_fill_manual(breaks = c(-1, 1), labels=c("Report \n To Police", "Not Report \n To Police"), values = c("#3148f7", "#35b5ff")) +
    facet_wrap(~both_infos) + 
    scale_alpha(range = c(0, 1)) +
    theme(panel.spacing = unit(4, "lines"),
          strip.text = element_text(size = 14),
          panel.spacing.x = unit(3, "cm")) +
    guides(fill=FALSE, group=FALSE, alpha=FALSE) +
  ggtitle('Responses to the Realistic Dilemma') +
  theme(plot.title = element_text(hjust = 0.5))
dilemma_decision_plot_real
```
## For the Philosophical Dilemma
```{r}
dilemma_decision_plot_philo <- ggplot()+
    geom_col(data=means_philo, aes(x=ingroup_norm, y=meanResponse, fill=factor(ingroup_norm)), width=vertical_bar_width, alpha=0.5) +
    geom_rect(data=props_philo, aes(xmin=xmin, ymin=ymin, xmax=xmax, ymax=ymax, group=ingroup_norm), fill = "#9c9d9e", alpha = 0.6) +
    scale_y_continuous(name="Preference", limits=c(1,6), oob = rescale_none, breaks = 1:6, labels=c("Definitely \n Diverting Train", "","","","", "Definitely \n Not Diverting Train")) + 
    scale_x_continuous(name = "Ingroup Norm", breaks=c(-1,1), labels=c("Divert Train", "Not Divert Train"))+
    scale_fill_manual(breaks = c(-1, 1), labels=c("Divert Train", "Not Divert Train"), values = c("#3148f7", "#35b5ff")) +
    facet_wrap(~both_infos) + 
    scale_alpha(range = c(0, 1)) +
    theme(panel.spacing = unit(4, "lines"),
          strip.text = element_text(size = 14),
          panel.spacing.x = unit(3, "cm")) +
    guides(fill=FALSE, group=FALSE, alpha=FALSE) +
  ggtitle('Responses to the Philosophical Dilemma') +
  theme(plot.title = element_text(hjust = 0.5))
dilemma_decision_plot_philo
```

# Plotting the Participants' Topic Choices
```{r, message=FALSE}
remaining_participants <- cleaned_data %>% pull(submission_id) %>% unique()
raw_data <- read_csv('../../Data/02_main/01_raw_main_data.csv')
raw_data_of_remaining_participants <- raw_data %>% filter(submission_id %in% remaining_participants)

topic_counts <- raw_data_of_remaining_participants %>% filter(trial_name == 'topic_choice') %>% group_by(response) %>% count()

topic_counts
```

```{r}
topic_choice_plot <- ggplot()+
  geom_col(data = topic_counts, aes(x = n, y = response), fill = '#6060FF')+
  ggtitle("Participants' Topic Choices")+
  theme(plot.title = element_text(hjust = 0.5))+
  scale_y_discrete(name="Topic", labels=c("Speed Limit", "Data Privacy","Gender Debate","Compulsory Vaccination","Migration", "Wealth Tax", "Cannabis Legalization", "Meat Consumption")) +
  theme(strip.text = element_text(size = 14))
  

topic_choice_plot
```
